+++ packages/nextjs-drupal-webform/package.json
@@ -1,0 +1,33 @@
+{
+  "name": "nextjs-drupal-webform",
+  "version": "1.0.0",
+  "private": true,
+  "license": "MIT",
+  "scripts": {
+    "build": "next build",
+    "dev": "next dev"
+  },
+  "dependencies": {
+    "@tailwindcss/typography": "^0.5.0",
+    "classnames": "^2.3.1",
+    "drupal-jsonapi-params": "^2.1.0",
+    "html-react-parser": "^1.4.8",
+    "next": "^12.2.0",
+    "next-acms": "^1.0.0-beta1",
+    "next-drupal": "^1.1.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "sharp": "^0.30.2"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.18.9",
+    "@types/node": "^18.0.6",
+    "@types/react": "^17.0.4",
+    "autoprefixer": "^10.4.2",
+    "eslint": "^8.20.0",
+    "eslint-config-next": "^12.2.3",
+    "postcss": "^8.4.5",
+    "tailwindcss": "^3.0.15",
+    "typescript": "^4.6.0"
+  }
+}
