+++ packages/nextjs-drupal-webform/pages/api/webform/submit.tsx
@@ -1,5 +1,5 @@
 import { NextApiRequest, NextApiResponse } from 'next';
-import { drupal } from '../../../lib/drupal';
+import { drupal } from '../../../src/drupal';
 
 export default async function handler(
   request: NextApiRequest,
