+++ starters/basic-starter/components/node--person.tsx
@@ -3,7 +3,7 @@ import Link from 'next/link';
 import { MediaImage } from 'components/media--image';
 import { FormattedText } from 'components/formatted-text';
 
-export function NodePerson({ node, additionalContent, ...props }) {
+export function NodePerson({ node, ...props }) {
   return (
     <article className="max-w-2xl px-6 py-10 mx-auto" {...props}>
       <div className="grid items-center justify-center gap-4 text-center md:text-left md:grid-cols-2">
